refactor(ai): moves withMetadata to generic test utils
