+++ packages/ai-commands/test/sql-util.ts
@@ -1,4 +1,3 @@
-import chalk from 'chalk'
 import { A_Expr, CreatePolicyStmt, Node, parseQuery } from 'libpg-query'
 
 export type PolicyInfo = {
@@ -168,46 +167,4 @@ export async function getPolicyInfo(createPolicyStatement: CreatePolicyStmt) {
   }
 
   return policyInfo
-}
-
-/**
- * Prints the provided metadata along with any assertion errors.
- * Works both synchronously and asynchronously.
- *
- * Useful for providing extra context for failed tests.
- */
-export function withMetadata<T extends void | Promise<void>>(
-  metadata: Record<string, string>,
-  fn: () => T
-): T {
-  /**
-   * Prepends metadata to an Error's stack trace.
-   */
-  function modifyError(err: unknown) {
-    if (err instanceof Error && err.stack) {
-      const formattedMetadata = Object.entries(metadata).map(
-        ([key, value]) => `${chalk.bold.dim(key)}:\n\n${chalk.green.dim(value)}`
-      )
-      err.stack = `${formattedMetadata.join('\n\n')}\n\n${err.stack}`
-    }
-
-    return err
-  }
-
-  // Execute the function and handle both
-  // synchronous or asynchronous scenarios
-  try {
-    const maybePromise = fn()
-
-    if (maybePromise instanceof Promise) {
-      return maybePromise.catch((err) => {
-        // Re-throw the error
-        throw modifyError(err)
-      }) as T
-    }
-    return maybePromise
-  } catch (err) {
-    // Re-throw the error
-    throw modifyError(err)
-  }
 }
